+++ lld/docs/ReleaseNotes.rst
@@ -24,7 +24,10 @@ Non-comprehensive list of changes in this release
 ELF Improvements
 ----------------
 
-* ...
+* ``-Bsymbolic -Bsymbolic-functions`` has been changed to behave the same as ``-Bsymbolic-functions``. This matches GNU ld.
+  (`D102461 <https://reviews.llvm.org/D102461>`_)
+* ``-Bno-symbolic`` has been added.
+  (`D102461 <https://reviews.llvm.org/D102461>`_)
 
 Breaking changes
 ----------------